requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.4
django-bootstrap4==3.0.1
django-guardian==2.4.0
django-tinymce==3.3.0
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
